Added more null checks added more junit renamed class for more intuitive term
